Refactor spectrogram scroller into modular package
